requirements.txt
transformers==4.39.1
peft==0.8.2
thop
datasets==2.17.1
wandb
tensorflow
packaging
gitpython
nltk
sentencepiece
matplotlib
pytorch-lightning==1.1.8
spacy 
scikit-learn
bert-score==0.3.9
tabulate
pandas
jsonlines
deepspeed==0.13.4
statsmodels
pytorch_toolbelt==0.8.0
torch==2.6.0
torchvision==0.21.0
accelerate==0.27.2